release用main.pyとdebug用main_debug.pyを分離
